Rename wrappers module to scenarios
