+++ baybe/simulation/__init__.py
@@ -16,8 +16,8 @@ applied context:
 """
 
 from baybe.simulation.core import simulate_experiment
+from baybe.simulation.scenarios import simulate_scenarios
 from baybe.simulation.transfer_learning import simulate_transfer_learning
-from baybe.simulation.wrappers import simulate_scenarios
 
 __all__ = [
     "simulate_experiment",
